+++ src/main/java/com/baedal/owner/adapter/in/web/controller/OwnerController.java
@@ -9,7 +9,7 @@ import com.baedal.owner.adapter.in.web.mapper.OwnerWebMapper;
 import com.baedal.owner.application.command.AddStoreCommand;
 import com.baedal.owner.application.command.LoginCommand;
 import com.baedal.owner.application.command.SignupCommand;
-import com.baedal.owner.application.port.in.OwnerSignUpUsecase;
+import com.baedal.owner.application.port.in.OwnerAuthenticateUsecase;
 import com.baedal.owner.application.port.in.OwnerUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -29,19 +29,19 @@ public class OwnerController {
 
   private final OwnerUseCase ownerUseCase;
 
-  private final OwnerSignUpUsecase signupUsecase;
+  private final OwnerAuthenticateUsecase authenticateUsecase;
 
   @PostMapping("/v0/login")
   public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest req) {
     LoginCommand.Request command = mapper.loginToCommand(req);
-    LoginCommand.Response commandResponse = ownerUseCase.login(command);
+    LoginCommand.Response commandResponse = authenticateUsecase.login(command);
     return ResponseEntity.ok(mapper.loginToResponse(commandResponse));
   }
 
   @PostMapping("/v0/signup")
   public ResponseEntity<SignUpResponse> signUp(@RequestBody SignUpRequest request) {
     SignupCommand.Request command = mapper.signupToCommand(request);
-    SignupCommand.Response commandResponse = signupUsecase.signUp(command);
+    SignupCommand.Response commandResponse = authenticateUsecase.signUp(command);
     return ResponseEntity.ok(mapper.signupToResponse(commandResponse));
   }
