+++ src/main/java/com/baedal/owner/application/port/in/OwnerAuthenticateUsecase.java
@@ -1,8 +1,11 @@
 package com.baedal.owner.application.port.in;
 
+import com.baedal.owner.application.command.LoginCommand;
 import com.baedal.owner.application.command.SignupCommand;
 
-public interface OwnerSignUpUsecase {
+public interface OwnerAuthenticateUsecase {
+
+  LoginCommand.Response login(LoginCommand.Request req);
 
   SignupCommand.Response signUp(SignupCommand.Request req);
 }
